requirements.txt
asgiref==3.6.0
Babel==2.11.0
beautifulsoup4==4.11.2
Django==4.1.6
django-bootstrap4==22.3
django-crispy-forms==2.0
django-extensions==3.2.1
django-money==3.0.0
django-mysql==4.8.0
django-phonenumber-field==7.0.2
django-widget-tweaks==1.4.12
mysqlclient==2.1.1
phonenumbers==8.13.5
Pillow==9.4.0
py-moneyed==2.0
python-decouple==3.7
python-gettext==4.1
pytz==2022.7.1
soupsieve==2.4
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.7
